Revert "move idConfig util to shared lib (#88)"
This reverts commit 13d07ef232a0aee93a77e0de7d58f5722b36ccaf.
