+++ gateway/package.json
@@ -48,7 +48,6 @@
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.2.2",
-    "shared": "../shared",
     "slugify": "^1.6.6",
     "sqlite3": "^5.1.7",
     "tsyringe": "^4.8.0",
