+++ services/catalog/db-scripts/seeds/seeders/category.seeder.ts
@@ -2,8 +2,8 @@ import {Repository} from "typeorm";
 import {CategoryEntity} from "../../../src/infrastructure/data/typeorm/entities/category.entity";
 // @ts-ignore
 import categoriesData from '../data/categories.json';
+import {UuidIdGenerator} from "../../../src/utils/id-generator.util";
 import slugify from "slugify";
-import {UuidIdGenerator} from "shared/lib/id-generator.util";
 
 export async function seedCategories(categoryRepo: Repository<CategoryEntity>): Promise<CategoryEntity[]> {
     const categoryEntities = [];
