+++ services/catalog/package.json
@@ -31,7 +31,6 @@
     "immutable": "^4.3.6",
     "joi": "^17.13.3",
     "reflect-metadata": "^0.2.2",
-    "shared": "../../shared",
     "slugify": "^1.6.6",
     "sqlite3": "^5.1.7",
     "tsyringe": "^4.8.0",
