+++ services/users/db-scripts/seeds/seeders/user.seeder.ts
@@ -1,7 +1,7 @@
 import {Repository} from "typeorm";
 // @ts-ignore
 import userData from '../data/users.json';
-import {UuidIdGenerator} from "../../../../../shared/src/id-generator.util";
+import {UuidIdGenerator} from "../../../src/utils/id-generator.util";
 import UserEntity from "../../../src/infrastructure/data/typeorm/entities/user.entity";
 
 export async function seedUsers(userEntityRepository: Repository<UserEntity>): Promise<UserEntity[]> {
