+++ services/users/src/core/services/user.service.ts
@@ -2,7 +2,7 @@ import User from "../models/user.model";
 import {DataSource, EntityManager} from "typeorm";
 import {UserRepository} from "../repositories/user.repository";
 import {inject, injectable} from "tsyringe";
-import {IdGenerator} from "../../../../../shared/src/id-generator.util";
+import {IdGenerator} from "../../utils/id-generator.util";
 
 export interface UserService {
     findById(id: string): Promise<User | null>;
